class-30/demo/pyproject.toml
[tool.poetry]
name = "hashtable"
version = "0.1.0"
description = ""
authors = ["JB-Tellez <jb.tellez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
codefellows = "^0.5.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
